+++ src/encoding_implementations/relative_mean_encoding.jl
@@ -48,6 +48,11 @@ struct RelativeMeanEncoding{R} <: Encoding
     end
 end
 
+# ----------------------------------------------------------------
+# Pretty printing (see /core/pretty_printing.jl).
+# ----------------------------------------------------------------
+hidefields(::Type{<:RelativeMeanEncoding}) = [:binencoder]
+
 function RelativeMeanEncoding(minval::Real, maxval::Real; n = 2)
     binencoder = RectangularBinEncoding(FixedRectangularBinning(0, 1, n + 1))
     return RelativeMeanEncoding(n, minval, maxval, binencoder)